get web tests to build
