+++ src/test/datascience/interactiveWindow.vscode.test.ts
@@ -21,7 +21,7 @@ import {
     submitFromPythonFileUsingCodeWatcher,
     waitForInteractiveWindow,
     waitForLastCellToComplete
-} from './helpers';
+} from './helpers.node';
 import {
     assertHasTextOutputInVSCode,
     clickOKForRestartPrompt,
